Extract FilteredSetsPanel out of sets route index
Keeps the route file to the route config and top-level component, with the panel/group-filtering logic moved to a dedicated component alongside its siblings in ArtistsTab.
